+++ app/Http/Controllers/ProductController.php
@@ -12,6 +12,19 @@ use Imagick;
 
 class ProductController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('permission:view products')->only(['index', 'show']);
+
+        $this->middleware('permission:create products')->only(['create', 'store']);
+
+        $this->middleware('permission:edit products')->only(['edit', 'update']);
+
+        $this->middleware('permission:delete products')->only(['destroy']);
+
+        $this->middleware('permission:upload product images')->only(['updateImage']);
+    }
+
     // generate sku
     private function generateSku()
     {
@@ -59,37 +72,27 @@ class ProductController extends Controller
         $imagePath = null;
 
         if ($request->hasFile('image')) {
-            $file = $request->file('image');
-            $filename = 'prd-' . time() . '.webp';
-
-            $path = storage_path('app/public/products/' . $filename);
-
-            $image = new Imagick($file->getRealPath());
-
-            $size = min($image->getImageWidth(), $image->getImageHeight());
-            $image->cropImage(
-                $size,
-                $size,
-                ($image->getImageWidth() - $size) / 2,
-                ($image->getImageHeight() - $size) / 2
-            );
-
-            $image->resizeImage(200, 200, Imagick::FILTER_LANCZOS, 1);
-
-            $image->setImageFormat('webp');
-            $image->setImageCompressionQuality(30);
-            $image->setOption('webp:method', '6');
-
-            $image->stripImage();
-
-            $image->writeImage($path);
-
-            $image->clear();
-            $image->destroy();
-
-            $imagePath = 'products/' . $filename;
+            if (auth()->user()->can('upload product images')) {
+                $file = $request->file('image');
+                $filename = 'prd-' . time() . '.webp';
+                $path = storage_path('app/public/products/' . $filename);
+                $image = new Imagick($file->getRealPath());
+                $size = min($image->getImageWidth(), $image->getImageHeight());
+                $image->cropImage($size, $size, ($image->getImageWidth() - $size) / 2, ($image->getImageHeight() - $size) / 2);
+                $image->resizeImage(200, 200, Imagick::FILTER_LANCZOS, 1);
+                $image->setImageFormat('webp');
+                $image->setImageCompressionQuality(30);
+                $image->setOption('webp:method', '6');
+                $image->stripImage();
+                $image->writeImage($path);
+                $image->clear();
+                $image->destroy();
+                $imagePath = 'products/' . $filename;
+            } else {
+                session()->flash('error', 'Gambar tidak diupload karena tidak punya izin');
+            }
         }
-
+        
         Product::create([
             'name' => $request->name,
             'sku' => $this->generateSku(),
